# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CryoET Catalog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/annotations/{sample_id}/{acquisition_id}/{reconstruction_alignment_id}/{annotation_id}/preview.png": {
      "get": {
        "tags": [
          "annotations"
        ],
        "summary": "Annotation Preview",
        "description": "Render the annotation MRC's center-XY slice as a PNG (1200px, 1\u201399%).\n\nReturns 404 for a missing row or path-outside-root, 422 for an annotation\nwith no ``.mrc`` artifact.",
        "operationId": "annotation_preview_annotations__sample_id___acquisition_id___reconstruction_alignment_id___annotation_id__preview_png_get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          },
          {
            "name": "acquisition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acquisition Id"
            }
          },
          {
            "name": "reconstruction_alignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Alignment Id"
            }
          },
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Annotation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/annotations/{sample_id}/{acquisition_id}/{reconstruction_alignment_id}/{annotation_id}/neuroglancer": {
      "post": {
        "tags": [
          "annotations"
        ],
        "summary": "Annotation Neuroglancer",
        "description": "Launch a Neuroglancer viewer over the annotation's ``.mrc`` volume.\n\nMirrors the tomogram launch route \u2014 same registry, same dev-side hostname\nrewrite on the frontend. 422 for an annotation with no ``.mrc`` artifact.",
        "operationId": "annotation_neuroglancer_annotations__sample_id___acquisition_id___reconstruction_alignment_id___annotation_id__neuroglancer_post",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          },
          {
            "name": "acquisition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acquisition Id"
            }
          },
          {
            "name": "reconstruction_alignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Alignment Id"
            }
          },
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Annotation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewerLaunchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ViewerLaunchOut": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ViewerLaunchOut",
        "description": "Response of a POST .../neuroglancer launch.\n\nFrontend rewrites the hostname to ``window.location.hostname`` before\nopening."
      },
...
    }
  }
}